% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datahandling_utils.R
\name{access_nested_list}
\alias{access_nested_list}
\title{Extracts values from nested list}
\usage{
access_nested_list(l, keys)
}
\arguments{
\item{l}{the list object}

\item{keys}{a vector of keys}
}
\value{
the extracted value
}
\description{
Extracts values from nested list
}
\examples{
lz = list(a1=list(b1 = "Streusalz",b2 = "Andreas Scheuer"), a2 = "Wurst")
lz = do.call(shiny::reactiveValues,lz)
access_nested_list(lz, c("a1","b2")) # should be "Andreas Scheuer"
}
