% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_data.R
\docType{data}
\name{BAMlogo_raster}
\alias{BAMlogo_raster}
\title{A raster representation of the BAM logo to be used in Reports.}
\format{
A raster representation of the BAM logo png file.
}
\source{
jan.lisec@bam.de
}
\usage{
data(BAMlogo_raster)
}
\description{
A raster representation of the BAM logo to be used in Reports.
}
\keyword{datasets}
