% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnc_c_filter_data.R
\name{c_filter_data}
\alias{c_filter_data}
\title{c_filter_data.}
\usage{
c_filter_data(x, c_apm)
}
\arguments{
\item{x}{The Cert data from an session R6 object.}

\item{c_apm}{The parameters of the current analyte from an session R6 object.}
}
\value{
A filtered data frame (S_flt, L_flt) without missing values and rounded to a specified precision.
}
\description{
\code{c_filter_data} will limit imported data from 'rv' for a specified analyte
according to parameters specified in 'apm'.
}
\details{
tbd.
}
\examples{
rv <- eCerto:::test_rv()
shiny::isolate(c_filter_data(
 x = getValue(rv, c("Certification","data")),
 c_apm = getValue(rv, c("General","apm"))[["Si"]]
))

}
