% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnc_checkHdata.R
\name{checkHdata}
\alias{checkHdata}
\title{fnc_checkHdata.}
\usage{
checkHdata(x)
}
\arguments{
\item{x}{The Hom data from an session R6 object.}
}
\value{
A data frame with at least columns 'analyte', 'H_type', 'Flasche' and 'value'.
}
\description{
\code{checkHdata} will check imported homogeneity data.
}
\details{
tbd.
}
\examples{
x <- eCerto:::test_homog()$data
eCerto:::checkHdata(x = x)
eCerto:::checkHdata(x = x[, -2])
}
\keyword{internal}
