% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_cert_data.R
\name{combine_cert_data}
\alias{combine_cert_data}
\title{combine_cert_data.}
\usage{
combine_cert_data(df_list = NULL, silent = FALSE)
}
\arguments{
\item{df_list}{list of already imported Excel tables.}

\item{silent}{Option to print or omit status messages.}
}
\value{
A dataframe.
}
\description{
\code{combine_cert_data} will post process data uploaded from Excel files of a certification trial.
}
\details{
not yet
}
