% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CRM001}
\alias{CRM001}
\title{An example set of data collected for a CRM.}
\format{
A list of length = 6 containing CRM test data.
}
\source{
jan.lisec@bam.de
}
\usage{
data(CRM001)
}
\description{
An example set of data collected for a CRM.
}
\keyword{datasets}
