% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datahandling_utils.R
\name{crop_dataframes}
\alias{crop_dataframes}
\title{Crops dataframe(s)}
\usage{
crop_dataframes(dfs, cols, rows)
}
\arguments{
\item{dfs}{list of dataframe(s)}

\item{cols}{columns as array, e.g. 1:3}

\item{rows}{rows as array, e.g. 5:8}
}
\value{
cropped list of data frames(s)
}
\description{
Crops dataframe(s)
}
\examples{
crop_dataframes(iris,2:3,5:6)
}
