% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_data.R
\docType{data}
\name{cvals_Dixon}
\alias{cvals_Dixon}
\title{Dixon critical values table.}
\format{
A data frame containing Dixon critical values with n in rows and alpha in cols.
}
\source{
\url{http://www.statistics4u.com/fundstat_eng/cc_outlier_tests_dixon.html}
}
\usage{
data(cvals_Dixon)
}
\description{
Dixon critical values table.
}
\keyword{datasets}
