% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_data.R
\docType{data}
\name{cvals_Grubbs2}
\alias{cvals_Grubbs2}
\title{Grubbs2 critical values table.}
\format{
A data frame containing critical values for Double Grubbs test with n in rows and alpha in cols.
}
\source{
\code{outliers} package and \url{https://link.springer.com/article/10.1007/s10182-011-0185-y}.
}
\usage{
data(cvals_Grubbs2)
}
\description{
Grubbs2 critical values table.
}
\keyword{datasets}
