% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datahandling_utils.R
\name{datahandling_utils}
\alias{datahandling_utils}
\alias{setValue}
\alias{getValue}
\alias{laboratory_dataframe}
\alias{access_nested_list}
\alias{set_nested_list}
\alias{load_sheetnames}
\alias{crop_dataframes}
\alias{data_of_godelement}
\alias{set_listUploadsource}
\alias{get_listUploadsource}
\alias{uploadsource_of_element}
\alias{roundMT}
\alias{pn}
\alias{update_reactivecell}
\title{setValue.}
\usage{
setValue(df, key, value)

getValue(df, key = NULL)

laboratory_dataframe(x)

access_nested_list(l, keys)

set_nested_list(l, keys, value)

load_sheetnames(filepath)

crop_dataframes(dfs, cols, rows)

data_of_godelement(l)

set_listUploadsource(rv, m, uploadsource)

get_listUploadsource(rv, m)

uploadsource_of_element(d)

roundMT(value, precision = NULL)

pn(n = NULL, p = 4L)

update_reactivecell(r, colname, analyterow = NULL, value)
}
\arguments{
\item{df}{An object of class R6.}

\item{key}{Key value within R6 object 'df'.}

\item{value}{value to be updated}

\item{x}{data frame with uploaded excel table}

\item{l}{list, which contains "data", but also e.g. "source"}

\item{keys}{a vector of keys}

\item{filepath}{the path to a single or multiple excel file(s)}

\item{dfs}{list of dataframe(s)}

\item{cols}{columns as array, e.g. 1:3}

\item{rows}{rows as array, e.g. 5:8}

\item{rv}{the list}

\item{m}{"Certification", etc.}

\item{uploadsource}{the source of upload}

\item{d}{Bin nicht sicher, ob diese Funktion sinnvoll ist.}

\item{precision}{precision value}

\item{n}{numeric vector}

\item{p}{precision after the decimal sign}

\item{r}{the reactive containing the data frame to be updated}

\item{colname}{name of the column}

\item{analyterow}{name of the analyte-row. If NULL, whole column is updated}
}
\value{
Nothing. The R6 object is updated automatically.

Value of 'key' from 'df'.

another data frame with extracted laboratory parameters

the extracted value

the names of sheets

cropped list of data frames(s)



nothing directly, but via rv





the rounded value

numbers formatted

nothing directly
}
\description{
General access to data object (so data object can maybe get changed without that much code edit)

Returns element. If 'key' is used, reactivity not working correctly.
Preferable way for calling \code{getValue(df)$key}, see example
}
\examples{
rv <- reactiveClass$new(init_rv())
setValue(rv, c("Certifications","data"), 5)
getValue(rv, c("Certifications","data")) # is 5?
datreturn <- ecerto:::test_datreturn()
shiny::isolate(ecerto::getValue(datreturn)$t_H)
lz = list(a1=list(b1 = "Streusalz",b2 = "Andreas Scheuer"), a2 = "Wurst")
lz = do.call(shiny::reactiveValues,lz)
access_nested_list(lz, c("a1","b2")) # should be "Andreas Scheuer"
ecerto:::crop_dataframes(iris,2:3,5:6)
if (interactive()) {
  rv = init_rv()
  set_listUploadsource(rv,"Certification","Excel")
}
roundMT(34.3434,3)
}
