% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploy_to_shinyio.R
\name{deploy_to_shinyio}
\alias{deploy_to_shinyio}
\title{deploy_to_shinyio.}
\usage{
deploy_to_shinyio(
  path = "C:/Users/jlisec/Rpackages/ShinyApps/eCerto",
  appName = c("TriboMat", "eCerto")[2]
)
}
\arguments{
\item{path}{Which path to use for the App.}

\item{appName}{Which shinyapp.io name is to be used.}
}
\value{
NULL.
}
\description{
\code{deploy_to_shinyio} will deploy_to_shinyio.
}
\details{
tbd.
}
\examples{
#'#ecerto:::deploy_to_shinyio()

}
