% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dixon.R
\name{Dixon}
\alias{Dixon}
\title{Title
BAMTool
Modul: Zertifizierung
Dixon}
\usage{
Dixon(lab_means = NULL)
}
\arguments{
\item{lab_means}{data.frame}
}
\value{

}
\description{
Title
BAMTool
Modul: Zertifizierung
Dixon
}
