% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/certification.R
\name{.CertificationServer}
\alias{.CertificationServer}
\title{Certification part}
\usage{
.CertificationServer(id, d, datreturn)
}
\arguments{
\item{id}{"namespace"}

\item{d}{d = reactive({rv$Certifications})}

\item{datreturn}{for storing the results}
}
\value{
nothing
}
\description{
Certification part
}
