% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/certification.R
\name{.CertificiationServer}
\alias{.CertificiationServer}
\title{Title}
\usage{
.CertificiationServer(id, d)
}
\arguments{
\item{id}{}

\item{d}{reactive Certification 'reactive({rv$Certifications})'}
}
\value{

}
\description{
Title
}
