% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlsx_modules.R
\name{.computation_final_data}
\alias{.computation_final_data}
\title{#' @rdname .xlsxinputServer
.xlsxinputUI = function(id) {
shiny::fileInput(
inputId = NS(id, "file"),
multiple = TRUE,
label = "Test-Upload (.xlsx format)",
accept = "xlsx"
)
}}
\usage{
.computation_final_data(id, a)
}
\description{
#' XLSX INPUT MODULE SERVER
#' Returns only the input from fileInput, which contains information about
#' the file path and their names
#'
#' @param id
#'
#' @return
#' @export
.xlsxinputServer = function(id) {
shiny::moduleServer(id, function(input, output, session) {\preformatted{# The selected file, if any
r = shiny::reactive(\{
  # If no file is selected, don't do anything
  shiny::req(input$file)
  input$file
\})
}

})
}
}
\section{TODO check if datafile is really an excel}{

}

\section{stopifnot(!is.reactivevalues(dat))}{

}

