% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlsx_modules.R
\name{.ExcelServer}
\alias{.ExcelServer}
\title{Excel Module
Receives sheet and path/name of datafiles and finally uploads
the Excel file with load_excelfiles(). Returns only raw uploaded
Excel tables.}
\usage{
.ExcelServer(id)
}
\arguments{
\item{id}{}
}
\value{
the raw and not-yet-cropped preview file
}
\description{
Excel Module
Receives sheet and path/name of datafiles and finally uploads
the Excel file with load_excelfiles(). Returns only raw uploaded
Excel tables.
}
