% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExcelUploadControl.R
\name{.ExcelUploadControl_UI}
\alias{.ExcelUploadControl_UI}
\title{ExcelUploadControl.}
\usage{
.ExcelUploadControl_UI(id)
}
\arguments{
\item{id}{Name when called as a module in a shiny app.}

\item{excelformat}{Selector for dataset type (reactive).}

\item{check}{Is current dataset slot empty (TRUE) or NULL (FALSE).}

\item{silent}{Option to print or omit status messages.}
}
\value{
A reactive dataframe.
}
\description{
\code{ExcelUploadControl} will provide a module to upload excel data files.
}
\details{
not yet
}
\examples{
shiny::shinyApp(
 ui = shiny::fluidPage(.ExcelUploadControl_UI(id = "test")),
 server = function(input, output, session) { out <- .ExcelUploadControl_Server(id = "test");  observeEvent(out(), {print(out())}) }
)

}
