% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlsx_modules.R
\name{.ExcelUploadControllServer}
\alias{.ExcelUploadControllServer}
\title{.ExcelUploadCntrServer}
\usage{
.ExcelUploadControllServer(id, c)
}
\arguments{
\item{id}{}

\item{c}{reactiveValues}
}
\value{

}
\description{
.ExcelUploadCntrServer
}
