% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lts_module.R
\name{.longtermstabilityUI}
\alias{.longtermstabilityUI}
\title{User Interface of LTS module}
\usage{
.longtermstabilityUI(id)
}
\arguments{
\item{id}{}
}
\value{

}
\description{
User Interface of LTS module
}
