% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlsx_modules.R
\name{.parameterServer}
\alias{.parameterServer}
\title{Title}
\usage{
.parameterServer(id, dat, excelformat)
}
\arguments{
\item{id}{}

\item{dat}{}

\item{excelformat}{}
}
\value{

}
\description{
Title
}
