% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlsx_modules.R
\name{.uploadTabsetsServer}
\alias{.uploadTabsetsServer}
\title{Module that receives and prints the preview, tests for invalidates and
also computes the final formatted dataset}
\usage{
.uploadTabsetsServer(id, excelformat, dat)
}
\arguments{
\item{id}{}

\item{excelformat}{}

\item{dat}{}
}
\value{

}
\description{
Module that receives and prints the preview, tests for invalidates and
also computes the final formatted dataset
}
