% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlsx_modules.R
\name{.xlsxinputServer}
\alias{.xlsxinputServer}
\title{XLSX INPUT MODULE SERVER
Returns only the input from fileInput, which contains information about
the file path and their names}
\usage{
.xlsxinputServer(id)
}
\arguments{
\item{id}{}
}
\value{

}
\description{
XLSX INPUT MODULE SERVER
Returns only the input from fileInput, which contains information about
the file path and their names
}
