% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnc_filter_data.R
\name{fnc_filter_data}
\alias{fnc_filter_data}
\title{fnc_filter_data.}
\usage{
fnc_filter_data(rv, an)
}
\arguments{
\item{rv}{The session R6 object.}

\item{an}{The name of the current analyte.}
}
\value{
A filtered data frame (S_flt, L_flt) without missing values.
}
\description{
\code{fnc_filter_data} will check parameters in 'apm' and limit imported data from 'rv' for a specified analyte.
}
\details{
tbd.
}
\examples{
rv <- reactiveClass$new(init_rv())
shiny::isolate({
eCerto::setValue(rv, c("Certification","data"),
 eCerto:::test_certification()[["data"]])
eCerto::setValue(rv, c("General","apm"),
 eCerto::init_apm(eCerto:::test_certification()[["data"]]))
fnc_filter_data(rv = rv, an="Si")
})

}
