% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnc_get_local_file.R
\name{fnc_get_local_file}
\alias{fnc_get_local_file}
\title{fnc_get_local_file.}
\usage{
fnc_get_local_file(x = NULL, pkg = "ecerto", copy_to_tempdir = TRUE)
}
\arguments{
\item{x}{Filename to search for without path but with extension.}

\item{pkg}{Package to seach system file in.}

\item{copy_to_tempdir}{If FALSE, file path is returned, if TRUE file is copied
to temp dir and this temp file path is returned.}
}
\value{
A file path.
}
\description{
\code{fnc_get_local_file} will provide a file path either from a
package system file (if available) or from www subdir (if online app).
}
\details{
Developing a shiny app as an R package in parallel leads to
difficulties on where to find specific files after installation (e.g. font
descriptions, markdown documents etc.). The function \code{fnc_get_local_file}
will check if a package of name pkg is installed and use system file to load,
assuming 'rmd' as the subdirectory to search in. If pgk is not available it
will provide a filepath using 'www' as the subdirectory for the shiny app
version.
}
\examples{
fnc_get_local_file(x="help_start.Rmd")
fnc_get_local_file(x="help_start.Rmd", copy_to_tempdir=FALSE)

}
