% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnc_load_RData.R
\name{fnc_load_RData}
\alias{fnc_load_RData}
\title{fnc_load_RData.}
\usage{
fnc_load_RData(x = NULL)
}
\arguments{
\item{x}{A list of values.}
}
\value{
A reactive values list 'rv'.
}
\description{
\code{fnc_load_RData} will check and convert values in an RData file and prepare the internal reactive values list 'rv'.
}
\details{
tbd.
}
\keyword{internal}
