% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnc_load_xlsx.R
\name{fnc_load_xlsx}
\alias{fnc_load_xlsx}
\title{fnc_load_xlsx.}
\usage{
fnc_load_xlsx(filepath, sheet, method = c("tidyxl", "openxlsx")[1], ...)
}
\arguments{
\item{filepath}{A list of imported Excel tables.}

\item{sheet}{Option to print or omit status messages.}

\item{method}{Either 'tidyxl' or 'openxlsx'.}

\item{...}{Further parameters to the read functions specified by method.}
}
\value{
A dataframe.
}
\description{
\code{fnc_load_xlsx} will handle upload of a single Excel file.
}
\details{
Function can handle reactive and non reactive parameters to be used in shiny apps.
}
\examples{
# test function with
x <- tempfile(fileext = ".xlsx")
openxlsx::write.xlsx(x=matrix(rnorm(9),ncol=3,dimnames=list(1:3,paste0("Header",1:3))), file=x)
fnc_load_xlsx(filepath=x, sheet=1)
fnc_load_xlsx(filepath=x, sheet=1, method="openxlsx")
fnc_load_xlsx(filepath="C:/not_existent.file", sheet=1)
fnc_load_xlsx(filepath=x, sheet=2)
x <- system.file(package="ecerto", "extdata", "EmptyExcel.xlsx")
fnc_load_xlsx(filepath=x, sheet=1, method="openxlsx")

}
