% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datahandling_utils.R
\name{get_listelem}
\alias{get_listelem}
\title{getter function for module element of the "god list"
modules so far are "Certifications, Homogeneity, Stability"}
\usage{
get_listelem(c, m)
}
\arguments{
\item{c}{"god list}

\item{m}{element to be fetched (e.g. "Certifications")}
}
\value{

}
\description{
getter function for module element of the "god list"
modules so far are "Certifications, Homogeneity, Stability"
}
