% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_utils.R
\name{get_listUploadsource}
\alias{get_listUploadsource}
\title{get source of upload for an element}
\usage{
get_listUploadsource(c, m)
}
\arguments{
\item{c}{}

\item{m}{}
}
\value{

}
\description{
get source of upload for an element
}
