% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datahandling_utils.R
\name{getValue}
\alias{getValue}
\title{Returns element. If 'key' is used, reactivity not working correctly.
Preferable way for calling \code{getValue(df)$key}, see example}
\usage{
getValue(df, key = NULL)
}
\arguments{
\item{df}{}

\item{key}{key, see notes}

\item{react}{should}
}
\value{

}
\description{
Returns element. If 'key' is used, reactivity not working correctly.
Preferable way for calling \code{getValue(df)$key}, see example
}
\examples{
datreturn = test_datreturn()
isolate(getValue(datreturn)$t_H)
}
