% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_the_user.R
\name{help_the_user}
\alias{help_the_user}
\title{Help Window: opens a modal with respective Help text for users}
\usage{
help_the_user(
  filename,
  format = c("rmd", "html", "rmd_with_link")[1],
  modal = TRUE
)
}
\arguments{
\item{filename}{name of the file as string (if necessary, containing also path)}

\item{format}{whether the file is html or rmd (default)}

\item{modal}{shall the modal window be opened (TRUE is default)}
}
\value{
returns the help text as HTML (currently produces errors when used)
}
\description{
Help Window: opens a modal with respective Help text for users
}
