% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_the_user_modal.R
\name{help_the_user_modal}
\alias{help_the_user_modal}
\title{Help Window: opens a modal with in memory rendering of simple Rmd files}
\usage{
help_the_user_modal(filename)
}
\arguments{
\item{filename}{name of the file as string (if necessary, containing also path)}
}
\value{
returns the help text as HTML (currently produces errors when used)
}
\description{
Help Window: opens a modal with in memory rendering of simple Rmd files
}
\keyword{internal}
