% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inits.R
\name{init_apm}
\alias{init_apm}
\title{Analyte Parameter List (apm)}
\usage{
init_apm(
  x = data.frame(ID = 1:20, analyte = gl(n = 2, k = 10, labels = c("A1", "A2")))
)
}
\arguments{
\item{x}{data frame of analyte names and sample IDs}
}
\value{
A empty list (if no data frame is provided) or a apm-List otherwise.
}
\description{
\code{init_apm} creates for each analyte the
parameter list. Each sublist contains information about the selected analyte
tab, and for each analyte the specified precision, the filtered sample id,
which sample ids are available to be filtered at all and, for completion,
the analyte name in case the list name fails
}
