% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inits.R
\name{init_datreturn}
\alias{init_datreturn}
\title{Initiate 'datreturn', which holds data returned from other modules meant for
temporary use}
\usage{
init_datreturn()
}
\value{
datreturn
}
\description{
Initiate 'datreturn', which holds data returned from other modules meant for
temporary use
}
