% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inits.R
\name{init_materialTabelle}
\alias{init_materialTabelle}
\title{Initializes material table for materialtabelle module}
\usage{
init_materialTabelle(analytes)
}
\arguments{
\item{analytes}{Character vector of analyte names.}
}
\value{
a data frame
}
\description{
Initializes material table for materialtabelle module
}
\examples{
analytes = c("Si","Ar")
matTab = init_materialTabelle(analytes)
}
