% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inits.R
\name{init_rv}
\alias{init_rv}
\title{Initialize permanent reactive values}
\usage{
init_rv()
}
\value{
a reactiveValues
}
\description{
\code{init_rv} initializes the main reactive value (rv) to store
the results from all modules. It therefore gets handed over multiple times.
In further programming, it should be considered to be replaced by a class
in OOP style. Note: If other modules besides Certification, Homogeneity and
Stability added, adapt the modules list
}
\examples{
rv = init_rv()
}
