% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datahandling_utils.R
\name{laboratory_dataframe}
\alias{laboratory_dataframe}
\title{creates long pivot table in laboratory style}
\usage{
laboratory_dataframe(x)
}
\arguments{
\item{x}{data frame with uploaded excel table}
}
\value{
another data frame with extracted laboratory parameters
}
\description{
creates long pivot table in laboratory style
}
