% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/certification_utils.R
\name{list2dataframe}
\alias{list2dataframe}
\title{transforms a list to a data frame}
\usage{
list2dataframe(l)
}
\arguments{
\item{l}{}
}
\value{

}
\description{
transforms a list to a data frame
}
