% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datahandling_utils.R
\name{listNames}
\alias{listNames}
\title{names of nested list elements, but ignore data.frame column names.
Refer to https://stackoverflow.com/q/68453593/6946122}
\usage{
listNames(l, maxDepth = 2)
}
\arguments{
\item{l}{nested list}

\item{maxDepth}{the maximum depth, the names of list should be returned}
}
\value{

}
\description{
names of nested list elements, but ignore data.frame column names.
Refer to https://stackoverflow.com/q/68453593/6946122
}
\examples{
a = list(
 b = list(df1 = data.frame(col = c(1, 2)), e = list(z = NULL)),
 c = NULL,
 df2 = data.frame(c12 = c(1, 2), c34 = c(3, 4)))
listNames(a,2) # [1] "b.df1" "b.e"   "c"     "df2"
}
