% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datahandling_utils.R
\name{load_sheetnames}
\alias{load_sheetnames}
\title{Loads names of Excel sheets}
\usage{
load_sheetnames(filepath)
}
\arguments{
\item{filepath}{the path to a single or multiple excel file(s)}
}
\value{
the names of sheets
}
\description{
Loads names of Excel sheets
}
