% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_data.R
\docType{data}
\name{LTS001}
\alias{LTS001}
\title{An example set of data collected for a LTS monitoring.}
\format{
A list of lists of length = 2 containing LTS test data.
}
\source{
jan.lisec@bam.de
}
\usage{
data(LTS001)
}
\description{
An example set of data collected for a LTS monitoring.
}
\keyword{datasets}
