% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m_analyte.R
\name{m_analyte}
\alias{m_analyte}
\alias{m_analyteUI}
\alias{m_analyteServer}
\title{Analyte-module}
\usage{
m_analyteUI(id)

m_analyteServer(id, apm, renewTabs, tablist)
}
\arguments{
\item{id}{Name when called as a module in a shiny app.}

\item{apm}{reactiveValues object, which gives available analytes, holds parameter, etc.}

\item{renewTabs}{The order, to delete and renew Tabs, for example when new data is uploaded}

\item{tablist}{list of current tabs (temporarily, see https://git.bam.de/fkress/ecerto/-/issues/46)}
}
\value{
the currently selected tab and Other parameter via apm reactiveValues()
}
\description{
\code{m_analyte} Module for organizing the analyte panel, which
fill automatically after analytes are available and gives the user the
opportunity to select analytes as well as precision and filter samples.
}
\details{
not yet
}
\examples{
if (interactive()) {
df <- data.frame("analyte"=gl(n = 2, k = 10, labels = c("A1","A2")))
shiny::shinyApp(
 ui = shiny::fluidPage(
 shinyjs::useShinyjs(),
   m_analyteUI(id = "test")
 ),
 server = function(input, output, session) {
   m_analyteServer(
     id = "test",
     apm = shiny::reactiveVal(analyte_parameter_list(df)),
     renewTab = reactiveVal(1),
     tablist = reactiveVal()
   )
 }
)
}

}
