% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyte_module.R
\name{m_analyteModuleUI}
\alias{m_analyteModuleUI}
\title{Analyte-module}
\usage{
m_analyteModuleUI(id)
}
\arguments{
\item{id}{Name when called as a module in a shiny app.}

\item{apm}{reactiveValues object, which gives available analytes, holds parameter, etc.}
}
\value{
nothing directly, works over apm parameter
}
\description{
\code{m_analyte} Module for organizing the analyte panel, which
fill automatically after analytes are available and gives the user the
opportunity to select analytes as well as precision and filter samples.
}
\details{
not yet
}
