% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m_stability_arrhenius.R
\name{m_arrheniusUI}
\alias{m_arrheniusUI}
\alias{m_arrheniusServer}
\title{m_stability_arrhenius.}
\usage{
m_arrheniusUI(id)

m_arrheniusServer(id, rv)
}
\arguments{
\item{id}{Name when called as a module in a shiny app.}

\item{rv}{The whole R6 object.}
}
\value{
A reactive indicating that the user wants to switch back
from arrhenius to simple view of S modul.
}
\description{
m_stability_arrhenius.
}
\examples{
if (interactive()) {
shiny::shinyApp(
 ui = shiny::fluidPage(
   m_arrheniusUI(id = "arrhenius")
 ),
 server = function(input, output, session) {
 rv <- eCerto:::test_rv()
 x <- eCerto:::test_Stability_Arrhenius()
 isolate(setValue(rv, c("Stability", "data"), x))
 out <- m_arrheniusServer(id = "arrhenius", rv = rv)
 shiny::observeEvent(out$switch, { print(out$switch) })
 }
)
}

}
