% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/certification.R
\name{m_CertificationUI}
\alias{m_CertificationUI}
\title{CERTIFICATION MODULE}
\usage{
m_CertificationUI(id)
}
\arguments{
\item{certification}{= reactive({rv$Certifications})}

\item{datreturn}{the session data object}
}
\value{
nothing directly, works over apm parameter
}
\description{
\code{m_CertificationServer} is the module for handling the
Certification part but also contains the materialtabelle (until further
changes).
}
\details{
not yet
}
