% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cert_Loaded.R
\name{m_CertLoadedUI}
\alias{m_CertLoadedUI}
\title{CERTIFICATION LOADED MODULE}
\usage{
m_CertLoadedUI(id)
}
\arguments{
\item{certification}{= reactive({rv$Certifications})}

\item{apm}{reactiveValues with analyte parameter}

\item{selected_tab}{currently selected tab}
}
\value{
nothing directly, works over apm parameter
}
\description{
\code{m_CertLoaded} is active as soon as Certification data was
uploaded.
}
\details{
not yet
}
