% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rdata_import.R
\name{m_RDataImport_UI}
\alias{m_RDataImport_UI}
\title{RDataImport.}
\usage{
m_RDataImport_UI(id)
}
\arguments{
\item{id}{Name when called as a module in a shiny app.}

\item{rv}{ReavtiveValues $$.}

\item{silent}{Option to print or omit status messages.}
}
\value{
A reactive dataframe.
}
\description{
\code{RDataImport} will provide a module to upload/backup Rdata files for certification trial data.
}
\details{
not yet
}
\examples{
shiny::shinyApp(
 ui = shiny::fluidPage(m_RDataImport_UI(id = "test")),
 server = function(input, output, session) { 
   out <- m_RDataImport_Server(id = "test");  
   observeEvent(out$Certifications$time_stamp, {
     print(out$Certifications$time_stamp)
   }) 
 }
)

}
