% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m_materialtabelle.R
\name{m_materialtabelleUI}
\alias{m_materialtabelleUI}
\alias{m_materialtabelleServer}
\title{m_materialtabelle.}
\usage{
m_materialtabelleUI(id)

m_materialtabelleServer(id, rv)
}
\arguments{
\item{id}{Name when called as a module in a shiny app.}

\item{rv}{eCerto R6 object, which includes a 'materialtabelle'.}
}
\value{
Nothing. Will update 'materialtabelle' in eCerto R6 object and trigger
an observer 'update_c_analyte' to be listened to externally.
}
\description{
\code{m_materialtabelle}
}
\details{
This module will show the reactive value 'materialtabelle' from the
eCerto R6 object in an editable table along with some action buttons.
}
\examples{
if (interactive()) {
shiny::shinyApp(
 ui = shiny::fluidPage(
   m_materialtabelleUI(id = "test")
 ),
 server = function(input, output, session) {
 rv <- eCerto:::test_rv()
 gargoyle::init("update_c_analyte")
 m_materialtabelleServer(id = "test", rv=rv)
 observeEvent(gargoyle::watch("update_c_analyte"), {print(rv$c_analyte)}, ignoreInit=TRUE)
 }
)
}

}
\keyword{internal}
