% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datahandling_utils.R
\name{merge_transfer}
\alias{merge_transfer}
\title{Merges a transferred column into the materialtabelle. If a value in the
column is exceeded by the transfer, it is overwritten. Otherwise, all other
values of the column and vector are combined.}
\usage{
merge_transfer(df, vec)
}
\arguments{
\item{df}{The data frame to transfer into}

\item{vec}{The transferred array (column) as dataframe}
}
\value{
The new data frame
}
\description{
Merges a transferred column into the materialtabelle. If a value in the
column is exceeded by the transfer, it is overwritten. Otherwise, all other
values of the column and vector are combined.
}
\examples{
df = data.frame(a = rep(0,4), b = c(0,0,10,0))
vec = data.frame(b = c(1,0,0,0))
merge_transfer(df = df, vec = vec)
}
