% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m_dataView.R
\name{mod_DataView}
\alias{mod_DataView}
\alias{m_DataViewUI}
\alias{m_DataViewServer}
\title{Data View Module}
\usage{
m_DataViewUI(id)

m_DataViewServer(id, dataset_flt, precision = shiny::reactiveVal(3))
}
\arguments{
\item{id}{Id when called in module.}

\item{dataset_flt}{Filtered data for particular measurements specified by user.}

\item{precision}{Precision to round the input values.}
}
\value{
kompakt data representation
}
\description{
Data View Module
}
\examples{
if (interactive()) {
shiny::shinyApp(
 ui = shiny::fluidPage(m_DataViewUI(id = "test")),
 server = function(input, output, session) {
   tmp <- reactiveVal(cbind(
     eCerto::test_certification()[["data"]],
     data.frame("File"="")
   ))
   m_DataViewServer(id = "test", dataset_flt = tmp)
 }
)
}

}
