% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m_materialtabelle.R
\name{mod_materialtabelle}
\alias{mod_materialtabelle}
\alias{m_materialtabelleUI}
\alias{m_materialtabelleServer}
\title{Modul "Materialtabelle"}
\usage{
m_materialtabelleUI(id)

m_materialtabelleServer(id, rv)
}
\arguments{
\item{id}{Name when called as a module in a shiny app.}

\item{rv}{The whole R6 object.}
}
\value{
The analyte name of the currently selected row of  mat_tab.
}
\description{
\code{m_materialtabelle}
}
\details{
not yet
}
\examples{
if (interactive()) {
shiny::shinyApp(
 ui = shiny::fluidPage(shinyalert::useShinyalert(), m_materialtabelleUI(id = "test")),
 server = function(input, output, session) {
 rv <- reactiveClass$new(init_rv())
 shiny::isolate(eCerto::setValue(rv, c("Certification","data"), eCerto:::test_certification()[["data"]]))
 shiny::isolate(eCerto::setValue(rv, c("General","apm"), eCerto::init_apm(eCerto:::test_certification()[["data"]])))
 shiny::isolate(eCerto::setValue(rv, c("General","materialtabelle"), eCerto::init_materialTabelle(names(eCerto::getValue(rv, c("General","apm"))))))
 out <- m_materialtabelleServer(id = "test", rv=rv)
 observeEvent(out(), {print(out())})
 }
)
}

}
