% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m_RData_export.R
\name{mod_RDataExport}
\alias{mod_RDataExport}
\alias{m_RDataExport_UI}
\alias{m_RDataexport_Server}
\title{Modul RData-Export}
\usage{
m_RDataExport_UI(id)

m_RDataexport_Server(id, rv)
}
\arguments{
\item{id}{Name when called as a module in a shiny app.}

\item{rv}{ReavtiveValues $$.}
}
\value{
rdata A reactive, but only for notifying the navbarpanel to change
}
\description{
Modul RData-Export
}
\details{
not yet
}
\examples{
if (interactive()) {

shiny::shinyApp(
 ui = shiny::fluidPage(m_RDataExport_UI(id = "test"), shiny::actionButton("insert","insert rv")),
 server = function(input, output, session) {
 rv_test <- eCerto::eCerto$new(init_rv())
   m_RDataexport_Server(id = "test", rv = rv_test)
   observeEvent(input$insert,{
   shiny::isolate({setValue(rv_test, c("Certification","data"), test_Certification_Excel()) })
    shiny::isolate({setValue(rv_test, c("General", "user"), "FK4") })
    shiny::isolate({set_uploadsource(rv_test, "Certification", uploadsource = "Excel") })
   })
 }
)
}
}
