% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m_report.R
\name{mod_report}
\alias{mod_report}
\alias{m_reportUI}
\alias{m_report_ui}
\alias{m_report_server}
\alias{m_reportServer}
\title{report}
\usage{
m_reportUI(id)

m_reportServer(id, rv)
}
\arguments{
\item{id}{module ID}

\item{rv}{the rv eCerto object}

\item{selected_tab}{which analyte-tab is currently selected}
}
\value{
nothing
}
\description{
report
}
\examples{
if (interactive()) {
shiny::shinyApp(
 ui = shiny::fluidPage(
   m_reportUI(id = "test")
 ),
 server = function(input, output, session) {
   rv <- eCerto::eCerto$new() # initiate persistent variables
   shiny::isolate({setValue(rv, c("General","study_id"), "Jan") })
   shiny::isolate({set_uploadsource(rv, "Certification", uploadsource = "Excel") })
   m_reportServer(
     id = "test",
     rv = rv
   )
 }
)
}

}
\keyword{internal}
