% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stability.R
\name{mod_Stability}
\alias{mod_Stability}
\alias{m_StabilityUI}
\title{STABILITY MODULE}
\usage{
m_StabilityUI(id)
}
\arguments{
\item{id}{}
}
\value{

}
\description{
STABILITY MODULE
}
\examples{
if (interactive()) {
shiny::shinyApp(
 ui = shiny::fluidPage(
   m_StabilityUI(id = "test")
 ),
 server = function(input, output, session) {
   rv <- ecerto::reactiveClass$new(ecerto::init_rv()) # initiate persistent variables
   isolate({ecerto::setValue(rv, c("Stability","data"), ecerto:::test_Stability_Excel() )})
   isolate({ecerto::setValue(rv, c("Stability","uploadsource"), "Excel")})
   datreturn <- ecerto:::test_datreturn()
   m_StabilityServer(
     id = "test",
     rv = rv,
     datreturn = datreturn
   )
 }
)
}

}
