% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m_start.R
\name{mod_Start}
\alias{mod_Start}
\alias{m_startUI}
\alias{m_startServer}
\title{start MODULE}
\usage{
m_startUI(id)

m_startServer(id, rv)
}
\arguments{
\item{id}{Name when called as a module in a shiny app.}
}
\value{
h_vals = The start data (not the transferred ones yet)
}
\description{
\code{m_start} is the module for eCerto startup.
}
\details{
Providing the backup/restore modules as well as the Excel upload.
}
\examples{
if (interactive()) {
shiny::shinyApp(
 ui = shiny::fluidPage(
   m_startUI(id = "test")
 ),
 server = function(input, output, session) {
   rv <- reactiveClass$new(init_rv()) # initiate persistent variables
   datreturn <- ecerto:::test_datreturn()
   m_startServer(id = "test", rv = rv)
 }
)
}

}
