% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lts_helper.R
\name{mondf}
\alias{mondf}
\title{mondf.}
\usage{
mondf(
  x = NULL,
  d_start = NULL,
  type = c("year", "mon", "day")[2],
  origin = "1900-01-01"
)
}
\arguments{
\item{x}{A vector of dates or character in format 'yyyy-mm-dd'.}

\item{d_start}{A specific start date (if NULL the minimum of x will be used).}

\item{type}{You may specify 'year' or 'day' instead of month here.}

\item{origin}{The origin used.}
}
\value{
number of months
}
\description{
Calculation of number of month since start date.
}
