% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistic_helper.R
\name{n_round_DIN1333}
\alias{n_round_DIN1333}
\title{n_round_DIN1333.}
\usage{
n_round_DIN1333(x)
}
\arguments{
\item{x}{A number.}
}
\description{
Determine the number of digits to round a uncertainty to depending on the position of the first
non-zero digit after the decimal.
}
\examples{
x <- c(0.011, 0.021, 0.0299999, 0.03, 0.031, 0.000299, 29.01, 3.01, 200, 300)
eCerto:::n_round_DIN1333(x)
}
\keyword{internal}
