% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/page_certification.R
\name{page_Certification}
\alias{page_Certification}
\alias{page_CertificationUI}
\alias{page_CertificationServer}
\title{Certification page}
\usage{
page_CertificationUI(id)

page_CertificationServer(id, rv, datreturn)
}
\arguments{
\item{id}{Name when called as a module in a shiny app.}

\item{rv}{The whole R6 object.}

\item{datreturn}{The session data object.}
}
\value{
nothing
}
\description{
\code{page_Certification} is the module for handling the
Certification part but also contains the materialtabelle (until further
changes).
}
\details{
not yet
}
\examples{
if (interactive()) {
shiny::shinyApp(
 ui = shiny::fluidPage(
   page_CertificationUI(id = "test")
 ),
 server = function(input, output, session) {
  rv <- reactiveClass$new(init_rv()) # initiate persistent variables
  shiny::isolate({setValue(rv, c("Certification","data"), ecerto:::test_Certification_Excel()) })
  shiny::isolate({set_uploadsource(rv, "Certification", uploadsource = "Excel") })
  datreturn <- reactiveClass$new(init_datreturn()) # initiate runtime variables

 page_CertificationServer(
     id = "test",
     rv = rv,
     datreturn = datreturn
   )
 }
)
}

}
