% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/page_stability.R
\name{page_Stability}
\alias{page_Stability}
\alias{page_StabilityUI}
\alias{page_StabilityServer}
\title{page_Stability}
\usage{
page_StabilityUI(id)

page_StabilityServer(id, rv, datreturn)
}
\arguments{
\item{id}{Id when called in module.}

\item{rv}{reactiveClass object.}

\item{datreturn}{reactiveClass object.}
}
\value{
Will return UI and Server logic for the stability page.
}
\description{
page_Stability
}
\examples{
if (interactive()) {
shiny::shinyApp(
 ui = shiny::fluidPage(
   page_StabilityUI(id = "test")
 ),
 server = function(input, output, session) {
   rv <- ecerto::reactiveClass$new(ecerto::init_rv()) # initiate persistent variables
   shiny::isolate({ecerto::setValue(rv, c("Stability","data"), ecerto:::test_Stability_Excel() )})
   shiny::isolate({ecerto::setValue(rv, c("Stability","uploadsource"), "Excel")})
   datreturn <- ecerto:::test_datreturn()
   page_StabilityServer(
     id = "test",
     rv = rv,
     datreturn = datreturn
   )
 }
)
}

}
