% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/page_start.R
\name{page_Start}
\alias{page_Start}
\alias{page_startUI}
\alias{page_startServer}
\title{Start-Page}
\usage{
page_startUI(id)

page_startServer(id, rv, tde)
}
\arguments{
\item{id}{Name when called as a module in a shiny app.}

\item{rv}{The gloobal R6 object.}

\item{tde}{test data environment.}
}
\value{
h_vals = The start data (not the transferred ones yet)
}
\description{
\code{page_start} is the module for eCerto startup.
}
\details{
Providing the backup/restore modules as well as the Excel upload.
}
\examples{
if (interactive()) {
shiny::shinyApp(
 ui = shiny::fluidPage(
   page_startUI(id = "test")
 ),
 server = function(input, output, session) {
   rv <- reactiveClass$new(init_rv()) # initiate persistent variables
   datreturn <- ecerto:::test_datreturn()
   page_startServer(id = "test", rv = rv)
 }
)
}

}
