% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lts_helper.R
\name{plot_lts_data}
\alias{plot_lts_data}
\title{plot_lts_data.}
\usage{
plot_lts_data(x = NULL, type = 1)
}
\arguments{
\item{x}{data.}

\item{type}{type of plot (see return).}
}
\value{
The plot function can be used to return only the calculated LTS value (in month) with type=0. If type=1 or 2 the normal or adjusted LTS plot will be computed additionally.
}
\description{
Plots for LTS module.
}
