% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datahandling_utils.R
\name{pn}
\alias{pn}
\title{format a number by rounding to a precision in same width as character using
scientific notation for numbers < precision and rounding to precision
otherwise}
\usage{
pn(n = NULL, p = 4L)
}
\arguments{
\item{n}{numeric vector}

\item{p}{precision after the decimal sign}
}
\value{
numbers formatted
}
\description{
format a number by rounding to a precision in same width as character using
scientific notation for numbers < precision and rounding to precision
otherwise
}
