% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnc_prepFigH1.R
\name{prepFigH1}
\alias{prepFigH1}
\title{fnc_prepFigH1.}
\usage{
prepFigH1(x, sa = NULL, prec = 4, xlab = "Flasche")
}
\arguments{
\item{x}{The Hom data from an session R6 object.}

\item{sa}{Selected analyte (name.type combination within x).}

\item{prec}{Precision value (for plot annotations).}
}
\value{
A data frame.
}
\description{
\code{prepFigH1} will perform statistics on imported homogeneity data.
}
\details{
tbd.
}
\examples{
x <- eCerto:::test_homog()$data
eCerto:::prepFigH1(x = x, sa = "Fe.radial")
}
\keyword{internal}
