% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnc_prepFigS1.R
\name{prepFigS1}
\alias{prepFigS1}
\title{fnc_prepFigS1.}
\usage{
prepFigS1(s, an, apm = NULL, U_Def = "2s", mt = NULL)
}
\arguments{
\item{s}{The imported stability data table.}

\item{an}{Name of the current analyte.}

\item{apm}{Analyte parameter list.}

\item{U_Def}{The deviation to be shown in the plot. Can be \code{U} if \code{mt} is present. Defaults to \verb{2s}.}

\item{mt}{materialtabelle to possibly contain data on \code{U_abs}, \code{cert_val} and \code{unit}.}
}
\value{
A list of length=2 containing measurement data ('val') and analyte definition ('def').
}
\description{
\code{prepFigS1} will prepare stability data for the lts plot function.
}
\details{
tbd.
}
\examples{
# s <- s_Data()
s <- eCerto:::test_Stability_Excel()
# an <- input$s_sel_analyte
# apm <- getValue(rv, c("General", "apm"))
apm <- list("Mn"=list("confirmed"=TRUE))
x_prep <- eCerto:::prepFigS1(s = s, an = "Mn", apm=apm)
eCerto:::plot_lts_data(x=x_prep)
# mt <- getValue(rv, c("General", "materialtabelle"))
mt <- data.frame("analyte"="Mn", "cert_val"=1, "U_abs"=1, "sd"=1, "unit"="unit")
x_prep <- eCerto:::prepFigS1(s = s, an = "Mn", apm=apm, U_Def="U", mt=mt)
eCerto:::plot_lts_data(x=x_prep)
}
\keyword{internal}
