% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnc_prepFigS2.R
\name{prepFigS2}
\alias{prepFigS2}
\title{fnc_prepFigS2.}
\usage{
prepFigS2(
  tmp,
  show_reference_point = TRUE,
  plot_nominal_scale = TRUE,
  plot_in_month = TRUE,
  plot_ln_relative = TRUE
)
}
\arguments{
\item{tmp}{The S data from an session R6 object.}

\item{show_reference_point}{show_reference_point.}

\item{plot_nominal_scale}{plot_nominal_scale.}

\item{plot_in_month}{plot_in_month.}

\item{plot_ln_relative}{plot_ln_relative.}
}
\value{
A data frame.
}
\description{
\code{prepFigS2} will generate a temp dependend set of plots.
}
\details{
tbd.
}
\examples{
x <- eCerto:::test_Stability_Arrhenius(3)
x$Value <- x$Value/mean(x$Value[x$time==0])
eCerto:::prepFigS2(tmp = x)
eCerto:::prepFigS2(tmp = x, show_reference_point = FALSE)
eCerto:::prepFigS2(tmp = x, plot_nominal_scale = FALSE)
eCerto:::prepFigS2(tmp = x, plot_in_month = FALSE)
eCerto:::prepFigS2(tmp = x, plot_ln_relative = FALSE)
}
\keyword{internal}
