% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnc_prepFigS3.R
\name{prepFigS3}
\alias{prepFigS3}
\title{fnc_prepFigS3.}
\usage{
prepFigS3(tab)
}
\arguments{
\item{tab}{The table containing data as basis for plot generation.}
}
\value{
A data frame.
}
\description{
\code{prepFigS3} will generate a temp dependend set of plots.
}
\details{
tbd.
}
\examples{
x <- data.frame(
  "dummy_name" = c("0", "4"),
  "1/K" = 1/(c(0, 4)+273.15),
  "log(-k_eff)" = c(1, 2),
  "CI_upper" = c(2, 3),
  "CI_lower" = c(0, 1),
  "log(k)_calc" = c(1, 2),
  check.names = FALSE
)
eCerto:::prepFigS3(tab = x)
}
\keyword{internal}
