% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnc_prepTabH1.R
\name{prepTabH1}
\alias{prepTabH1}
\title{fnc_prepTabH1.}
\usage{
prepTabH1(x)
}
\arguments{
\item{x}{The Hom data from an session R6 object.}
}
\value{
A data frame.
}
\description{
\code{prepTabH1} will perform statistics on imported homogeneity
data.
}
\details{
tbd.
}
\examples{
x <- eCerto:::test_homog()$data
eCerto:::prepTabH1(x = x)
}
\keyword{internal}
