% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactiveClass.R
\name{reactiveClass}
\alias{reactiveClass}
\title{reactive class}
\description{
builds a class, which allows only restricted access to the contained
reactiveValues. elements should be accessed via \code{\link[=getValue]{getValue()}}.
\enumerate{
\item structure of #'reactiveValues is clear from the beginning (no function like "addVariable"
should exist!)
\item functions to calculate, e.g., the mean can be implemented
}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{reactiveClass$new()}}
\item \href{#method-set}{\code{reactiveClass$set()}}
\item \href{#method-get}{\code{reactiveClass$get()}}
\item \href{#method-names}{\code{reactiveClass$names()}}
\item \href{#method-clone}{\code{reactiveClass$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{reactiveClass$new(rv)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set"></a>}}
\if{latex}{\out{\hypertarget{method-set}{}}}
\subsection{Method \code{set()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{reactiveClass$set(field = NULL, value)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get"></a>}}
\if{latex}{\out{\hypertarget{method-get}{}}}
\subsection{Method \code{get()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{reactiveClass$get(field = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-names"></a>}}
\if{latex}{\out{\hypertarget{method-names}{}}}
\subsection{Method \code{names()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{reactiveClass$names()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{reactiveClass$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
