% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lts_helper.R
\name{read_lts_input}
\alias{read_lts_input}
\title{read_lts_input.}
\usage{
read_lts_input(file = NULL, simplify = FALSE)
}
\arguments{
\item{file}{Path to Excel File}

\item{simplify}{Try to simplify list of imported Excel files into data.frame}
}
\value{
Excel
}
\description{
Read Excel function for Long-term-stability data.
}
