% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistic_helper.R
\name{round_up}
\alias{round_up}
\title{round_up.}
\usage{
round_up(x, n = 0)
}
\arguments{
\item{x}{A number.}

\item{n}{precision after decimal.}
}
\description{
Round always up.
}
\examples{
x <- c(0.011, -0.021, 0.0299999, 0.03, 0.031, 0.000299, 29.01, 3.01, 200, -300)
eCerto:::round_up(x, n=2)
eCerto:::round_up(x=x, n=eCerto:::n_round_DIN1333(x))
}
\keyword{internal}
