% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistic_helper.R
\name{roundK}
\alias{roundK}
\title{roundK.}
\usage{
roundK(x, n = 0)
}
\arguments{
\item{x}{A number.}

\item{n}{precision after decimal.}
}
\description{
Round .5 always up (non ISO standard).
}
\examples{
x <- c(-2.5,2.5,3.5)
round(x)
eCerto:::roundK(x)
}
\keyword{internal}
