% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datahandling_utils.R
\name{roundMT}
\alias{roundMT}
\title{Rounds material table.}
\usage{
roundMT(value, precision = NULL)
}
\arguments{
\item{value}{the value to be rounded}

\item{precision}{precision value}
}
\value{
the rounded value
}
\description{
Rounds material table.
}
\examples{
roundMT(34.3434,3)
}
