% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_utils.R
\name{save_link}
\alias{save_link}
\title{save_link.}
\usage{
save_link(text, link)
}
\arguments{
\item{text}{text.}

\item{link}{link.}
}
\description{
Provide save link tag
}
\examples{
save_link("KS", "https://rdrr.io/r/stats/ks.test.html")
}
