% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_utils.R
\name{set_listelem}
\alias{set_listelem}
\title{setter function for an element in the list}
\usage{
set_listelem(c, m, elem)
}
\arguments{
\item{c}{}

\item{m}{}

\item{elem}{}
}
\value{

}
\description{
setter function for an element in the list
}
