% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datahandling_utils.R
\name{set_listUploadsource}
\alias{set_listUploadsource}
\title{#' getter function for module element of the "god list"
#' modules so far are "Certifications, Homogeneity, Stability"
#'
#' @param c "god list
#' @param m element to be fetched (e.g. "Certifications")
#'
#' @return
#' @export
get_listelem = function(c, m) {
.Deprecated("getValue")
data_of_godelement(c[\link{m}])
}
#' setter function for an element in the "god list"
#'
#' @param rv "god list"
#' @param m element to be fed (e.g. "Certifications")
#' @param dat data to be inserted
#'
#' @return
#' @export
set_listelem = function(rv, m, dat) {if(!is.null(get_listelem(c,m))) {warning(paste0(m, " in list is not null"))return(NULL)}

.Deprecated("setValue")}
\usage{
set_listUploadsource(rv, m, uploadsource)
}
\arguments{
\item{rv}{the list}

\item{m}{one of "Certification","Homogeneity", "Stability"}

\item{uploadsource}{the source of upload}
}
\value{
nothing directly, but via rv
}
\description{
if(is.reactive(dat)) {
rv$set(c(m,"data"),isolate(dat()))
# c[\link{m}][\link{"data"}] = isolate(dat())
} else {
rv$set(c(m,"data"),dat)
# c[\link{m}][\link{"data"}] = dat
}
}
\details{

}
\examples{
## Not run: 
rv = init_rv()
set_listUploadsource(rv,"Certification","Excel")
}
