% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datahandling_utils.R
\name{set_nested_list}
\alias{set_nested_list}
\title{\code{\link[=similar to access_nested_list]{similar to access_nested_list()}}][ecerto::access_nested_list]}
\usage{
set_nested_list(l, keys, value)
}
\description{
\code{\link[=similar to access_nested_list]{similar to access_nested_list()}}][ecerto::access_nested_list]
}
