% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datahandling_utils.R
\name{setValue}
\alias{setValue}
\title{General access to data object (so data object can maybe get changed without that much code edit)}
\usage{
setValue(df, key, value)
}
\arguments{
\item{df}{the data frame (e.g. a R6 object)}

\item{key}{key(s)}

\item{value}{value to set}
}
\value{
nothing
}
\description{
General access to data object (so data object can maybe get changed without that much code edit)
}
