% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datahandling_utils.R
\name{show_view}
\alias{show_view}
\title{show_view}
\usage{
show_view(rv)
}
\arguments{
\item{rv}{The eCerto R6 object.}
}
\value{
A character vector of panels to be shown or more precisely the parent list names
which contain a sub item 'show' that is set to TRUE.
}
\description{
Show View Returns a list of panels, which are marked to be shown in the
accordingly used RData from previous analysis. This is currently not evaluated
but could be useful in the future. Keep for now and don't delete,
}
\examples{
rv <- eCerto::eCerto$new(init_rv()) # initiate persistent variables
shiny::isolate({setValue(rv, c("Certification_processing","CertValPlot","show"),TRUE) })
print(eCerto:::show_view(rv))
shiny::isolate({setValue(rv, c("Certification_processing","mstats","show"),TRUE) })
print(eCerto:::show_view(rv))
}
\keyword{internal}
