% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sliderupdate.R
\name{sliderupdate}
\alias{sliderupdate}
\title{updates slider in excel tabset (not in use anymore)}
\usage{
sliderupdate(session, dat)
}
\arguments{
\item{session}{}

\item{dat}{}
}
\value{

}
\description{
updates slider in excel tabset (not in use anymore)
}
