% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnc_styleTabH1.R
\name{styleTabH1}
\alias{styleTabH1}
\title{fnc_styleTabH1.}
\usage{
styleTabH1(x, mt = NULL, prec = NULL, output = c("df", "dt")[1], cr = 1)
}
\arguments{
\item{x}{The Hom data from an session R6 object.}

\item{mt}{The mt from an session R6 object.}

\item{prec}{The precision of all analytes from x (names vector).}

\item{output}{Return either the dataframe with styling information in columns or the corresponding datatable object.}

\item{cr}{Current row selected (relevant if output = 'dt').}
}
\value{
A data frame or a datatable object depending on parameter 'output'.
}
\description{
\code{styleTabH1} will style Tab.H1 for pretty output.
}
\details{
tbd.
}
\examples{
x <- eCerto:::prepTabH1(x = eCerto:::test_homog()$data)
x
eCerto:::styleTabH1(x = x)
mt <- data.frame("analyte"="Fe")
eCerto:::styleTabH1(x = x, mt = mt)
prec <- unlist(list("Fe"=2))
eCerto:::styleTabH1(x = x, prec = prec)
eCerto:::styleTabH1(x = x, output = "dt", prec = prec)
}
\keyword{internal}
