% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnc_styleTabS1.R
\name{styleTabS1}
\alias{styleTabS1}
\title{fnc_styleTabS1.}
\usage{
styleTabS1(x, mt = NULL, sr = 1)
}
\arguments{
\item{x}{The S data from an session R6 object.}

\item{mt}{The mt from an session R6 object.}

\item{sr}{Currently selected row of table.}
}
\value{
A datatable object.
}
\description{
\code{styleTabS1} will style Tab.S1 for pretty output.
}
\details{
tbd.
}
\keyword{internal}
