% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_helper.R
\name{test_homog}
\alias{test_homog}
\title{test_homog.}
\usage{
test_homog()
}
\description{
test_homog.
}
\details{
The homogeneity test data are hard coded here using \code{structure} to
potentially identify changes in the excel importer function. When tested,
the excel importer will read the file "Homog_test.xlsx" from the folder
\code{ìnst/extdata} and compare the result with the return value of this function.
}
\keyword{internal}
