% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestPlot.R
\name{TestPlot}
\alias{TestPlot}
\title{BAMTool
Modul: Zertifizierung
Test Plot
x : data table}
\usage{
TestPlot(data = NULL)
}
\arguments{
\item{data}{data.frame with ID, Lab, analyte, replicate, value, unit, S_flt, L_flt}
}
\description{
BAMTool
Modul: Zertifizierung
Test Plot
x : data table
}
