% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datahandling_utils.R
\name{to_startPage}
\alias{to_startPage}
\title{to_startPage.}
\usage{
to_startPage(session, value = "Certification")
}
\arguments{
\item{session}{session.}

\item{value}{value.}
}
\description{
to_startPage.
}
\keyword{Page}
\keyword{Start}
\keyword{internal}
\keyword{switch}
\keyword{to}
