% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datahandling_utils.R
\name{update_reactivecell}
\alias{update_reactivecell}
\title{Update single or multiple cells of the final materialtabelle (formerly cert_vals)
with new values}
\usage{
update_reactivecell(r, colname, analyterow = NULL, value)
}
\arguments{
\item{r}{the reactive containing the data frame to be updated}

\item{colname}{name of the column}

\item{analyterow}{name of the analyte-row. If NULL, whole column is updated}

\item{value}{value to be updated}
}
\value{

}
\description{
Update single or multiple cells of the final materialtabelle (formerly cert_vals)
with new values
}
