% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datahandling_utils.R
\name{uploadsource_of_element}
\alias{uploadsource_of_element}
\title{Returns source of upload for an element}
\usage{
uploadsource_of_element(d)
}
\arguments{
\item{d}{}
}
\value{

}
\description{
Returns source of upload for an element
}
