% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{xlsxload_uploadTabsetsServer}
\alias{xlsxload_uploadTabsetsServer}
\title{Title}
\usage{
xlsxload_uploadTabsetsServer(id, excelformat, dat)
}
\arguments{
\item{id}{}

\item{excelformat}{}

\item{dat}{}
}
\value{

}
\description{
Title
}
