% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inits.R
\name{analyte_parameter_list}
\alias{analyte_parameter_list}
\title{Analyte Parameter List (apm)}
\usage{
analyte_parameter_list(certification = NULL)
}
\arguments{
\item{certification}{data frame of certification data}
}
\value{
reactiveValues
}
\description{
\code{analyte_parameter_list} creates for each analyte the
parameter list. Each sublist contains information about he selected analyte
tab, and for each analyte the wanted precision, the filtered sample id,
which sample ids are available to be filtered at all and, for completion,
the analyte name in case the list name fails
}
\examples{
apm = analyte_parameter_list(data_of_godelement(rv$Certifications))
}
