% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnc_CertValPlot.R
\name{CertValPlot}
\alias{CertValPlot}
\title{CertValPlot.}
\usage{
CertValPlot(data = NULL, annotate_id = FALSE)
}
\arguments{
\item{data}{data.frame containing columns 'value', 'Lab' and 'Filter'.}
}
\value{
A specific type of boxplot.
}
\description{
\code{CertValPlot} will generate a certified values plot.
}
\details{
not yet
}
