% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{CertValPlot}
\alias{CertValPlot}
\title{creates Certificate Validation Boxplot
BAMTool
Modul: Zertifizierung
Certified Value Plot
data : data table}
\usage{
CertValPlot(data = NULL)
}
\arguments{
\item{data}{}
}
\value{

}
\description{
creates Certificate Validation Boxplot
BAMTool
Modul: Zertifizierung
Certified Value Plot
data : data table
}
