% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/certification_utils.R
\name{data_of_godelement}
\alias{data_of_godelement}
\title{Returns the "data" element of the current "god list" element}
\usage{
data_of_godelement(d)
}
\arguments{
\item{d}{}
}
\value{

}
\description{
Returns the "data" element of the current "god list" element
}
