% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_utils.R
\name{datahandling_utils}
\alias{datahandling_utils}
\alias{setValue}
\alias{getValue}
\title{setValue.}
\usage{
setValue(df, key, value)

getValue(df, key = NULL)
}
\arguments{
\item{df}{An object of class R6.}

\item{key}{Key value within R6 object 'df'.}

\item{value}{Value to set.}
}
\value{
Nothing. The R6 object is updated automatically.

Value of 'key' from 'df'.
}
\description{
General access to data object (so data object can maybe get changed without that much code edit)

Returns element. If 'key' is used, reactivity not working correctly.
Preferable way for calling \code{getValue(df, key)}, see example
}
\examples{
# Only run examples in interactive R sessions
if (interactive()) {
  rv <- eCerto$new(init_rv())
  setValue(rv, c("Certification","data"), 5)
  getValue(rv, c("Certification","data")) # is 5?
  setValue(rv, c("General","user"),"Franz")
  getValue(rv, c("General","user"))
}
}
