% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datahandling_utils.R
\name{datahandling_utils}
\alias{datahandling_utils}
\alias{setValue}
\alias{getValue}
\alias{laboratory_dataframe}
\alias{load_sheetnames}
\alias{set_uploadsource}
\alias{roundMT}
\alias{pn}
\alias{update_reactivecell}
\title{setValue.}
\usage{
setValue(df, key, value)

getValue(df, key = NULL)

laboratory_dataframe(x)

load_sheetnames(filepath)

set_uploadsource(rv, m, uploadsource)

roundMT(value, precision = NULL)

pn(n = NULL, p = 4L)

update_reactivecell(r, colname, analyterow = NULL, value)
}
\arguments{
\item{df}{An object of class R6.}

\item{key}{Key value within R6 object 'df'.}

\item{value}{value to be updated}

\item{x}{data frame with uploaded excel table}

\item{filepath}{the path to a single or multiple excel file(s)}

\item{rv}{the list}

\item{m}{one of "Certification","Homogeneity", "Stability"}

\item{uploadsource}{the source of upload}

\item{precision}{precision value}

\item{n}{numeric vector}

\item{p}{requested precision after the decimal sign}

\item{r}{the reactive containing the data frame to be updated}

\item{colname}{name of the column}

\item{analyterow}{name of the analyte-row. If NULL, whole column is updated}
}
\value{
Nothing. The R6 object is updated automatically.

Value of 'key' from 'df'.

another data frame with extracted laboratory parameters

the names of sheets

nothing directly, but via rv

the rounded value

numbers formatted

nothing directly
}
\description{
General access to data object (so data object can maybe get changed without that much code edit)

Returns element. If 'key' is used, reactivity not working correctly.
Preferable way for calling \code{getValue(df, key)}, see example
}
\examples{
rv <- reactiveClass$new(init_rv())
# Only run examples in interactive R sessions
if (interactive()) {
 setValue(rv, c("Certification","data"), 5)
 getValue(rv, c("Certification","data")) # is 5?
 setValue(rv, c("General","user"),"Franz")
 getValue(rv, c("General","user"))
}
if (interactive()) {
  rv = init_rv()
  set_uploadsource(rv,"Certification","Excel")
}
roundMT(34.3434,3)
pn(n=c(1.23456, NA, 0, 0.00001, -0.00001))
}
