% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploy_to_shinyio.R
\name{deploy_to_shinyio}
\alias{deploy_to_shinyio}
\title{deploy_to_shinyio.}
\usage{
deploy_to_shinyio(
  path_app = "C:/Users/jlisec/Rpackages/ShinyApps/eCerto",
  path_pkg = "C:/Users/jlisec/Rpackages/Rpackage_eCerto/ecerto",
  appName = c("test", "eCerto")[1]
)
}
\arguments{
\item{path_app}{Which path to use for the App.}

\item{path_pkg}{Which path to use for the App.}

\item{appName}{Either shinyapp.io name to deploy to or NULL to omit deployment step.}
}
\value{
NULL.
}
\description{
\code{deploy_to_shinyio} will deploy_to_shinyio.
}
\details{
use \code{rsconnect::showLogs(appName = "test")} to check for errors
}
\examples{
\dontrun{ecerto:::deploy_to_shinyio(appName=NULL)}

}
\keyword{internal}
