% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/certification.R
\name{.CertificationServer}
\alias{.CertificationServer}
\title{Certification part}
\usage{
.CertificationServer(id, d, datreturn)
}
\arguments{
\item{id}{"namespace"}

\item{d}{reactive with uploaded and pre-formatted certification data.
Contains  $data (ID, Lab, analyte, replicate, value, unit, ...) and
$uploadsource (Excel, RData, etc.)}

\item{datreturn}{for storing the results}
}
\value{
nothing
}
\description{
Certification part
}
