% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlsx_modules.R
\name{.ExcelUploadControllServer}
\alias{.ExcelUploadControllServer}
\title{Here you can select the Ecerto-Mode (Certification, Homogeneity, Stability)
and, until now, contains the Upload button}
\usage{
.ExcelUploadControllServer(id, c)
}
\arguments{
\item{id}{}

\item{c}{reactiveValues}
}
\value{

}
\description{
Here you can select the Ecerto-Mode (Certification, Homogeneity, Stability)
and, until now, contains the Upload button
}
