% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlsx_modules.R
\name{.uploadTabsetsServer}
\alias{.uploadTabsetsServer}
\title{Title}
\usage{
.uploadTabsetsServer(id, excelformat, dat)
}
\arguments{
\item{id}{}

\item{excelformat}{}

\item{dat}{}
}
\value{

}
\description{
Title
}
