% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactiveClass.R
\name{eCerto_R6Class}
\alias{eCerto_R6Class}
\alias{eCerto}
\title{A reactive class based on an R6 object.}
\description{
Builds a class, which allows only restricted access to the
contained 'reactiveValues'. Elements should be accessed via \code{\link[=getValue]{getValue()}}.
Possible advantages are that (1) structure of 'reactiveValues' is clear
from the beginning (no function like "addVariable" should exist!) and that
(2) functions to calculate the mean or plot current data can be implemented
here directly.#'
}
\examples{
if(interactive()) {
rv <- shiny::reactiveValues(a=1)
shiny::observeEvent(rv$a, { message("rv$a changed:", rv$a) })
shiny:::flushReact()
rv$a <- 2
shiny:::flushReact()
test <- eCerto$new(eCerto:::init_rv())
eCerto::getValue(test, c("Certification","data"))
shiny::observeEvent(eCerto::getValue(test, "Certification")$data, {
  message("Certification$data changed:", eCerto::getValue(test, "Certification")$data)
})
eCerto::setValue(test, c("Certification","data"), 5)
eCerto::getValue(test, c("Certification","data"))
shiny:::flushReact()
}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-eCerto-new}{\code{eCerto$new()}}
\item \href{#method-eCerto-get}{\code{eCerto$get()}}
\item \href{#method-eCerto-set}{\code{eCerto$set()}}
\item \href{#method-eCerto-c_plot}{\code{eCerto$c_plot()}}
\item \href{#method-eCerto-clone}{\code{eCerto$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-eCerto-new"></a>}}
\if{latex}{\out{\hypertarget{method-eCerto-new}{}}}
\subsection{Method \code{new()}}{
Write the (reactive) value of element 'keys' from list 'l'.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{eCerto$new(rv)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{rv}}{'reactiveValues' object.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new 'eCerto' object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-eCerto-get"></a>}}
\if{latex}{\out{\hypertarget{method-eCerto-get}{}}}
\subsection{Method \code{get()}}{
Read the value of field element of R6 object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{eCerto$get(keys = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{keys}}{Name of list element.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Current value of field.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-eCerto-set"></a>}}
\if{latex}{\out{\hypertarget{method-eCerto-set}{}}}
\subsection{Method \code{set()}}{
Write the value to field element of R6 object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{eCerto$set(keys = NULL, value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{keys}}{Name of list element.}

\item{\code{value}}{New value.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
New value of field (invisible).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-eCerto-c_plot"></a>}}
\if{latex}{\out{\hypertarget{method-eCerto-c_plot}{}}}
\subsection{Method \code{c_plot()}}{
Plot the certification data either provided by the user or from the private slot of self.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{eCerto$c_plot(data, annotate_id = FALSE, filename_labels = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{data.frame containing columns 'value', 'Lab' and 'L_flt'.}

\item{\code{annotate_id}}{T/F to overlay the plot with ID as text if column 'ID' is present.}

\item{\code{filename_labels}}{T/F to use imported file names as labels on x-axes.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A plot.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-eCerto-clone"></a>}}
\if{latex}{\out{\hypertarget{method-eCerto-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{eCerto$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
