% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnc_get_local_file.R
\name{fnc_get_local_file}
\alias{fnc_get_local_file}
\title{fnc_get_local_file.}
\usage{
fnc_get_local_file(x = NULL, copy_to_tempdir = TRUE, fsep = .Platform$file.sep)
}
\arguments{
\item{x}{Filename to search for without path but with extension.}

\item{copy_to_tempdir}{If TRUE file is copied to a tempdir location.}

\item{fsep}{The system path seperator.
to temp dir and this temp file path is returned.}
}
\value{
A file path (either of source or of a temp file depending on option 'copy_to_tempdir').
}
\description{
\code{fnc_get_local_file} will provide a file path either from a
package system file (if available) or from www subdir (if online app).
}
\details{
Developing a shiny app as an R package in parallel leads to
difficulties on where to find specific files after installation (e.g. font
descriptions, markdown documents etc.). The function \code{fnc_get_local_file}
will check if a package 'of name pkg'ecerto' is installed and use system file to load.
If ecerto is not available it will search in the current wd.
}
\examples{
fnc_get_local_file(x="help_start.Rmd")
fnc_get_local_file(x="help_start.Rmd", copy_to_tempdir=FALSE)

}
