% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datahandling_utils.R
\name{get_listUploadsource}
\alias{get_listUploadsource}
\title{get source of upload for an element}
\usage{
get_listUploadsource(rv, m)
}
\arguments{
\item{rv}{the list}

\item{m}{"Certification", etc.}
}
\value{

}
\description{
get source of upload for an element
}
