% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnc_get_local_file.R
\name{get_local_file}
\alias{get_local_file}
\title{get_local_file.}
\usage{
get_local_file(x = NULL)
}
\arguments{
\item{x}{File pattern to search for.}
}
\value{
A file path.
}
\description{
\code{get_local_file} will provide a full file path using the
currently specified 'www' folder specified as a recourcePaths.
}
\details{
Developing a shiny app as an R package in parallel leads to difficulties
on where to find specific files after installation (e.g. font descriptions,
markdown documents etc.). The function \code{get_local_file} will search
a specfied pattern in the current recourcePaths for 'www'.
}
\keyword{internal}
