% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_the_user.R
\name{help_the_user}
\alias{help_the_user}
\title{Help Window: opens a modal with respective Help text for users}
\usage{
help_the_user(filename, format = "rmd", modal = TRUE)
}
\arguments{
\item{filename}{name of the file as string (if necessary, containing also path)}
}
\value{

}
\description{
Help Window: opens a modal with respective Help text for users
}
\examples{


}
