% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inits.R
\name{init_apm}
\alias{init_apm}
\title{Analyte Parameter List (apm)}
\usage{
init_apm(x)
}
\arguments{
\item{x}{Analyte data frame containing at least columns \code{ID}, \code{analyte} and \code{Lab}.}
}
\value{
The analyte parameter list (apm) including all individually settable options.
}
\description{
\code{init_apm} creates for each analyte the
parameter list. Each sublist contains information about the selected analyte
tab, and for each analyte the specified precision, the filtered sample id,
which sample ids are available to be filtered at all and, for completion,
the analyte name in case the list name fails
}
