% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inits.R
\name{init_apm}
\alias{init_apm}
\title{Analyte Parameter List (apm)}
\usage{
init_apm(certification = NULL)
}
\arguments{
\item{certification}{data frame of certification data}
}
\value{
reactiveValues
}
\description{
\code{init_apm} creates for each analyte the
parameter list. Each sublist contains information about the selected analyte
tab, and for each analyte the specified precision, the filtered sample id,
which sample ids are available to be filtered at all and, for completion,
the analyte name in case the list name fails
}
\examples{
apm <- shiny::reactiveVal()
df <- data.frame("analyte"=gl(n = 2, k = 10, labels = c("A1","A2")))
apm(init_apm(df))
apm_tmp <- shiny::isolate(apm())
apm_tmp[["A1"]]$confirmed = TRUE
apm(apm_tmp)
shiny::isolate(apm()[["A1"]]$confirmed) # TRUE

}
