% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inits.R
\name{init_datreturn}
\alias{init_datreturn}
\title{Initiates a reactiveValues object used for session-only and
module-overarching content}
\usage{
init_datreturn()
}
\value{
datreturn
}
\description{
Initiates a reactiveValues object used for session-only and
module-overarching content
}
