% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inits.R
\name{init_datreturn}
\alias{init_datreturn}
\title{init_datreturn.}
\usage{
init_datreturn()
}
\value{
datreturn
}
\description{
\code{init_datreturn} Initiates a reactiveValues object used for
session-only and module-to-materialtabelle content
}
