% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inits.R
\name{init_materialtabelle}
\alias{init_materialtabelle}
\title{Initializes material table for m_materialtabelle}
\usage{
init_materialtabelle(analytes)
}
\arguments{
\item{analytes}{Character vector of analyte names.}
}
\value{
a data frame
}
\description{
Initializes material table for m_materialtabelle
}
\examples{
matTab <- init_materialtabelle(analytes = c("Si","Ar"))
}
