% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lts_module.R
\name{longtermstability}
\alias{longtermstability}
\alias{.longtermstabilityUI}
\alias{longtermstabilityUI}
\alias{longtermstabilityServer}
\alias{.longtermstabilityServer}
\title{longtermstabilityServer.}
\usage{
.longtermstabilityUI(id)

.longtermstabilityServer(id)
}
\arguments{
\item{id}{Name when called as a module in a shiny app.}
}
\value{

}
\description{
Modul for LongTermStability Monitoring as currently (2021) used by Carsten Prinz.
}
