% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/materialtabelle.R
\name{m_materialtabelleUI}
\alias{m_materialtabelleUI}
\title{Modul "Materialtabelle"}
\usage{
m_materialtabelleUI(id)
}
\arguments{
\item{id}{Name when called as a module in a shiny app.}

\item{rDtataUpload}{if uploaded via RData - reactive({rv$Certifications$materialtabelle})}

\item{datreturn}{the session data (R6) object}
}
\value{
nothing
}
\description{
\code{m_materialtabelle}
}
\details{
not yet
}
