% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m_materialtabelle.R
\name{m_materialtabelleUI}
\alias{m_materialtabelleUI}
\alias{m_materialtabelleServer}
\title{m_materialtabelle.}
\usage{
m_materialtabelleUI(id)

m_materialtabelleServer(id, rv)
}
\arguments{
\item{id}{Name when called as a module in a shiny app.}

\item{rv}{The whole R6 object.}
}
\value{
The analyte name of the currently selected row of  mat_tab.
}
\description{
\code{m_materialtabelle}
}
\details{
not yet
}
\examples{
if (interactive()) {
shiny::shinyApp(
 ui = shiny::fluidPage(
   m_materialtabelleUI(id = "test")
 ),
 server = function(input, output, session) {
 rv <- eCerto:::test_rv()
 out <- m_materialtabelleServer(id = "test", rv=rv)
 observeEvent(out(), {print(out())})
 }
)
}

}
