% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/certification.R
\name{mod_Certification}
\alias{mod_Certification}
\alias{m_CertificationUI}
\alias{m_CertificationServer}
\title{Certification.}
\usage{
m_CertificationUI(id)

m_CertificationServer(id, rv, apm.input, datreturn)
}
\arguments{
\item{id}{Name when called as a module in a shiny app.}

\item{apm.input}{analyteParameterList, when uploaded from RData (reactive)}

\item{datreturn}{the session data object}

\item{certification}{reactive({getValue(rv,"Certifications")})}
}
\value{
analyte parameter
}
\description{
\code{m_CertificationServer} is the module for handling the
Certification part but also contains the materialtabelle (until further
changes).
}
\details{
not yet
}
\examples{
if (interactive()) {
rv <- ecerto::init_rv()
datreturn <- ecerto::init_datreturn()
shiny::shinyApp(
 ui = shiny::fluidPage(
   m_CertificationUI(id = "test")
 ),
 server = function(input, output, session) {
   m_CertificationServer(
     id = "test",
     certification = shiny::reactive({rv$Certifications}),
     datreturn = datreturn
   )
 }
)
}

}
