% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/certification.R
\name{mod_Certification}
\alias{mod_Certification}
\alias{m_CertificationUI}
\alias{m_CertificationServer}
\title{Certification.}
\usage{
m_CertificationUI(id)

m_CertificationServer(id, certification, datreturn)
}
\arguments{
\item{id}{Name when called as a module in a shiny app.}

\item{certification}{reactive({rv$Certifications})}

\item{datreturn}{the session data object}
}
\value{
nothing directly, works over apm parameter
}
\description{
\code{m_CertificationServer} is the module for handling the
Certification part but also contains the materialtabelle (until further
changes).
}
\details{
not yet
}
