% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m_certification.R
\name{mod_Certification}
\alias{mod_Certification}
\alias{m_CertificationUI}
\alias{m_CertificationServer}
\title{Certification.}
\usage{
m_CertificationUI(id)

m_CertificationServer(id, rv, datreturn)
}
\arguments{
\item{id}{Name when called as a module in a shiny app.}

\item{rv}{the whole R6 object}

\item{datreturn}{the session data object}
}
\value{
nothing
}
\description{
\code{m_CertificationServer} is the module for handling the
Certification part but also contains the materialtabelle (until further
changes).
}
\details{
not yet
}
\examples{
if (interactive()) {
shiny::shinyApp(
 ui = shiny::fluidPage(
   m_CertificationUI(id = "test")
 ),
 server = function(input, output, session) {
  rv <- reactiveClass$new(init_rv()) # initiate persistent variables
  shiny::isolate({setValue(rv, c("Certification","data"), ecerto:::test_Certification_Excel()) })
  shiny::isolate({set_uploadsource(rv, "Certification", uploadsource = "Excel") })
  datreturn <- reactiveClass$new(init_datreturn()) # initiate runtime variables

 m_CertificationServer(
     id = "test",
     rv = rv,
     datreturn = datreturn
   )
 }
)
}

}
