% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cert_Loaded.R
\name{mod_CertLoaded}
\alias{mod_CertLoaded}
\alias{m_CertLoadedUI}
\alias{m_CertLoadedServer}
\title{CERTIFICATION LOADED MODULE}
\usage{
m_CertLoadedUI(id)

m_CertLoadedServer(id, rv, apm, selected_tab)
}
\arguments{
\item{id}{Name when called as a module in a shiny app.}

\item{rv}{}

\item{apm}{reactiveVal with analyte parameter}

\item{selected_tab}{currently selected tab, for example "Si" or "Cu" etc.}
}
\value{
dat, which is data frame with ID, Lab, analyte, replicate, value, unit, S_flt, L_flt
}
\description{
\code{m_CertLoaded} is active as soon as Certification data was
uploaded.
}
\details{
additional returns are handled via apm
}
