% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m_dataView.R
\name{mod_DataView}
\alias{mod_DataView}
\alias{m_DataViewUI}
\alias{m_DataViewServer}
\title{Data View Module}
\usage{
m_DataViewUI(id)

m_DataViewServer(id, dataset_flt, current_apm)
}
\arguments{
\item{id}{Id when called in module.}

\item{dataset_flt}{filtered data for particular measurements specified by user}

\item{current_apm}{current analyte-parameter-list for selected analyte}
}
\value{
Will return UI and Server logic for the stability tab.
}
\description{
Data View Module
}
\examples{
if (interactive()) {
shiny::shinyApp(
 ui = shiny::fluidPage(
   m_DataViewUI(id = "test")
 ),
 server = function(input, output, session) {
 datasetflt = reactiveVal(structure(
 list(
   ID = c(1L,10L,18,25), Lab = factor(c("L1","L1","L2","L2")), 
   analyte = rep("Si",4), replicate = c(1,2,1,2),
    value = runif(4, 0, 1),unit = rep("0.05",4),
    File = rep("Ergebnisblatt_BAM-M321_Aleris_Duffel_m.xlsx",4),
    S_flt = rep(FALSE,4),L_flt = rep(FALSE, 4),
    row.names = c(1L,10L,18,25),class = "data.frame")))
    
   m_DataViewServer(
     id = "test",
     dataset_flt = datasetflt,
     current_apm  = reactiveVal(list(precision = 3))
   )
 }
)
}

}
