% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataView.R
\name{mod_DataView}
\alias{mod_DataView}
\alias{m_DataViewUI}
\alias{m_DataViewServer}
\title{Data View Module}
\usage{
m_DataViewUI(id)

m_DataViewServer(id, dataset_flt, current_apm)
}
\arguments{
\item{id}{Id when called in module.}

\item{current_apm}{current analyte-parameter-list for selected analyte}

\item{dat}{filtered data for particular measurements specified by user}
}
\value{
Will return UI and Server logic for the stability tab.
}
\description{
Data View Module
}
\examples{
if (interactive()) {
shiny::shinyApp(
 ui = shiny::fluidPage(
   m_DataViewUI(id = "test")
 ),
 server = function(input, output, session) {
   m_DataViewServer(
     id = "test"
   )
 }
)
}

}
