% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homogeneity.R
\name{mod_Homogeneity}
\alias{mod_Homogeneity}
\alias{m_HomogeneityUI}
\alias{m_HomogeneityServer}
\title{HOMOGENEITY MODULE}
\usage{
m_HomogeneityUI(id)

m_HomogeneityServer(id, homog, cert)
}
\arguments{
\item{id}{Name when called as a module in a shiny app.}

\item{homog}{= the Homogeneity table shiny::reactive({getValue(rv,"Homogeneity")})}

\item{cert}{= shiny::reactive({getValue(rv,"Certifications")})}
}
\value{
h_vals = The Homogeneity data (not the transferred ones yet)
}
\description{
\code{m_Homogeneity} is the module for handling Homogeneity Data
}
\details{
not yet
}
