% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/materialtabelle.R
\name{mod_materialtabelle}
\alias{mod_materialtabelle}
\alias{m_materialtabelleUI}
\alias{m_materialtabelleServer}
\title{Modul "Materialtabelle"}
\usage{
m_materialtabelleUI(id)

m_materialtabelleServer(id, rdataUpload, datreturn)
}
\arguments{
\item{id}{Name when called as a module in a shiny app.}

\item{rdataUpload}{if uploaded via RData - reactive({rv$Certifications$materialtabelle})}

\item{datreturn}{the session data (R6) object}
}
\value{
nothing
}
\description{
\code{m_materialtabelle}
}
\details{
not yet
}
\examples{
if (interactive()) {
shiny::shinyApp(
 ui = shiny::fluidPage(m_materialtabelleUI(id = "test")),
 server = function(input, output, session) {
 fpath <- system.file("extdata", "SR3_Fe_v26chs.RData", package="ecerto")
 load(fpath)
 rdataUpload <- reactive({res[[1]][["cert_vals"]]})
 datreturn <- ecerto:::test_datreturn()
 m_materialtabelleServer(id = "test", rdataUpload=rdataUpload, datreturn=datreturn);
  #observeEvent(out(), {print(out())})
 }
)
}

}
