% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m_RData_export.R
\name{mod_RDataExport}
\alias{mod_RDataExport}
\alias{m_RDataExport_UI}
\alias{m_RDataexport_Server}
\title{Modul RData-Export}
\usage{
m_RDataExport_UI(id)

m_RDataexport_Server(id, rv, silent = FALSE)
}
\arguments{
\item{id}{Name when called as a module in a shiny app.}

\item{rv}{ReavtiveValues $$.}

\item{silent}{Option to print or omit status messages.}
}
\value{
rdata A reactive, but only for notifying the navbarpanel to change
}
\description{
Modul RData-Export
}
\details{
not yet
}
\examples{
if (interactive()) {
   
shiny::shinyApp(
 ui = shiny::fluidPage(m_RDataExport_UI(id = "test")),
 server = function(input, output, session) {
 rv_test <- reactiveClass$new(init_rv())
    shiny::isolate({setValue(rv_test, c("Certification","data"), test_Certification_Excel()) })
    shiny::isolate({setValue(rv_test, c("General", "user"), "FK4") })
    shiny::isolate({set_uploadsource(rv_test, "Certification", uploadsource = "Excel") })
   m_RDataexport_Server(id = "test", rv = rv_test)
 }
)
}
}
