% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m_report.R
\name{mod_report}
\alias{mod_report}
\alias{m_report_ui}
\alias{m_report_server}
\title{report}
\usage{
m_report_ui(id)

m_report_server(id, rv, selected_tab)
}
\arguments{
\item{id}{module ID}

\item{rv}{the rv-reactiveClass object}

\item{selected_tab}{which analyte-tab is currently selected}
}
\value{
nothing
}
\description{
report
}
\examples{
if (interactive()) {
shiny::shinyApp(
 ui = shiny::fluidPage(
   m_report_ui(id = "test")
 ),
 server = function(input, output, session) {
   rv <- reactiveClass$new(init_rv()) # initiate persistent variables
   shiny::isolate({setValue(rv, c("General","study_id"), "Jan") })
   shiny::isolate({setValue(rv, c("Certification","data"), eCerto:::test_Certification_Excel()) })
   shiny::isolate({set_uploadsource(rv, "Certification", uploadsource = "Excel") })
   m_report_server(
     id = "test",
     rv = rv,
     selected_tab = reactiveVal("Si")
   )
 }
)
}

}
