% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lts_helper.R
\name{mondf}
\alias{mondf}
\title{mondf.}
\usage{
mondf(d_start, d_end)
}
\arguments{
\item{d_start}{A start date}

\item{d_end}{An end date}
}
\value{

}
\description{
Calculation of number of month between start and end date.
}
