% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/page_certification.R
\name{page_Certification}
\alias{page_Certification}
\alias{page_CertificationUI}
\alias{page_CertificationServer}
\title{Certification page}
\usage{
page_CertificationUI(id)

page_CertificationServer(id, rv)
}
\arguments{
\item{id}{Name when called as a module in a shiny app.}

\item{rv}{The whole R6 object.}
}
\value{
nothing
}
\description{
\code{page_Certification} is the module for handling the
Certification part but also contains the materialtabelle (until further
changes).
}
\details{
not yet
}
\examples{
if (interactive()) {
shiny::shinyApp(
 ui = shiny::fluidPage(
   shinyjs::useShinyjs(),
   page_CertificationUI(id = "test")
 ),
 server = function(input, output, session) {
  rv <- eCerto:::test_rv()
 page_CertificationServer(id = "test", rv = rv)
 }
)
}

}
