% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/page_homogeneity.R
\name{page_Homogeneity}
\alias{page_Homogeneity}
\alias{page_HomogeneityUI}
\alias{page_HomogeneityServer}
\title{Homogeneity page}
\usage{
page_HomogeneityUI(id)

page_HomogeneityServer(id, rv)
}
\arguments{
\item{id}{Name when called as a module in a shiny app.}

\item{rv}{The session R6 object.}
}
\description{
\code{page_Homogeneity} is the module for handling Homogeneity Data
}
\details{
not yet
}
\examples{
if (interactive()) {
shiny::shinyApp(
 ui = shiny::fluidPage(
   page_HomogeneityUI(id = "test")
 ),
 server = function(input, output, session) {
   rv <- eCerto:::test_rv()
   mt <- isolate(eCerto::getValue(rv, c("General","materialtabelle")))
   attr(mt, "col_code") <- data.frame("ID"="U","Name"="U")
   isolate(eCerto::setValue(rv, c("General","materialtabelle"), mt))
   isolate(eCerto::setValue(rv, "Homogeneity", eCerto:::test_homog()))
   page_HomogeneityServer(
     id = "test",
     rv = rv
   )
 }
)
}

}
