% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lts_helper.R
\name{plot_lts_data}
\alias{plot_lts_data}
\title{plot_lts_data.}
\usage{
plot_lts_data(x = NULL, type = 1)
}
\arguments{
\item{x}{data}

\item{type}{type of plot (either 1 oder 2)}
}
\value{

}
\description{
Plots for LTS module.
}
