% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m_Rdata_import.R
\name{RDataImport}
\alias{RDataImport}
\alias{m_RDataImport_UI}
\alias{m_RDataImport_Server}
\title{RDataImport.}
\usage{
m_RDataImport_UI(id)

m_RDataImport_Server(id, modules, uploadsources, silent = FALSE)
}
\arguments{
\item{id}{Name when called as a module in a shiny app.}

\item{modules}{c("Certification","Homogeneity","Stability").}

\item{uploadsources}{contains which of the \code{modules} has which
uploadsource or \code{NULL}. For example if Certification has been
uploaded, the argument would look like list("Certification" =
"Excel","Homogeneity"=NULL,"stability"=NULL).}

\item{silent}{Option to print or omit status messages.}
}
\value{
rdata A reactive, but only for notifying the navbarpanel to change
}
\description{
\code{RDataImport} will provide a module to upload/backup Rdata
files for certification trial data.
}
\details{
not yet
}
\examples{
if (interactive()) {
shiny::shinyApp(
 ui = shiny::fluidPage(ecerto::m_RDataImport_UI(id = "test")),
 server = function(input, output, session) {
   ecerto::m_RDataImport_Server(
     id = "test",
     modules = reactiveVal(c("Ceritification","Stability","Homogeneity")),
     uploadsource = reactiveVal("Excel")
     )
 }
)}

}
