% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lts_module.R
\name{read_lts_input}
\alias{read_lts_input}
\title{Read Excel for Long term stability}
\usage{
read_lts_input(file = NULL, simplify = FALSE)
}
\arguments{
\item{file}{}

\item{simplify}{}
}
\value{

}
\description{
Read Excel for Long term stability
}
