% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datahandling_utils.R
\name{roundMT}
\alias{roundMT}
\title{Rounds material table.
Currently without}
\usage{
roundMT(value, precision = NULL)
}
\arguments{
\item{value}{}

\item{precision}{}
}
\value{

}
\description{
Rounds material table.
Currently without
}
