% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datahandling_utils.R
\name{set_listelem}
\alias{set_listelem}
\title{setter function for an element in the "god list"}
\usage{
set_listelem(c, m, dat)
}
\arguments{
\item{c}{"god list}

\item{m}{element to be fed (e.g. "Certifications")}

\item{dat}{data}
}
\value{

}
\description{
setter function for an element in the "god list"
}
