% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datahandling_utils.R
\name{show_view}
\alias{show_view}
\title{Show View Returns a list of panels, which are marked to be shown in the
accordingly used RData from previous analysis}
\usage{
show_view(rv)
}
\arguments{
\item{rv}{the R6 reactiveValues object}
}
\value{
a list of panels to be shown
}
\description{
Show View Returns a list of panels, which are marked to be shown in the
accordingly used RData from previous analysis
}
\examples{
rv <- reactiveClass$new(init_rv()) # initiate persistent variables
shiny::isolate({setValue(rv, c("Certification_processing","CertValPlot","show"),TRUE) })
print(show_view(rv))
}
