% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnc_styleTabH1.R
\name{styleTabH1}
\alias{styleTabH1}
\title{fnc_styleTabH1.}
\usage{
styleTabH1(x, mt = NULL, apm = NULL)
}
\arguments{
\item{x}{The Hom data from an session R6 object.}

\item{mt}{The mt from an session R6 object.}

\item{apm}{The apm from an session R6 object.}
}
\value{
A data frame.
}
\description{
\code{styleTabH1} will style Tab.H1 for pretty output.
}
\details{
tbd.
}
\examples{
x <- eCerto:::prepTabH1(x = eCerto:::test_homog()$data)
x
eCerto:::styleTabH1(x = x)
mt <- data.frame("analyte"="Fe")
eCerto:::styleTabH1(x = x, mt = mt)
apm <- list("Fe"=list("precision"=2))
eCerto:::styleTabH1(x = x, apm = apm)
}
\keyword{internal}
