% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnc_styleTabH2.R
\name{styleTabH2}
\alias{styleTabH2}
\title{fnc_styleTabH2.}
\usage{
styleTabH2(x, precision = 4)
}
\arguments{
\item{x}{data.frame.}

\item{precision}{Rounding precision.}
}
\value{
A datatable object.
}
\description{
\code{styleTabH2} will style Tab.H2 for pretty output.
}
\details{
tbd.
}
\examples{
x <- data.frame("Item" = LETTERS[1:3], "mean" = 1:3, "sd" = rnorm(3), "n" = 4:6)
eCerto:::styleTabH2(x = x)
eCerto:::styleTabH2(x = x, precision = 2)
}
\keyword{internal}
